Separate mapping utils into files by entity
