+++ src/controllers/enrollmentReports/mapUtils/fundingTime.ts
@@ -1,0 +1,25 @@
+import { FundingSource } from '../../../../client/src/shared/models';
+import { FUNDING_SOURCE_TIMES } from '../../../../client/src/shared/constants';
+
+/**
+ * Leverage funding source -> time -> format mappings from FUNDING_SOURCE_TIMES
+ * to determine the valid funding time entered by the user.
+ */
+export const mapFundingTime = (
+  value: number | string | undefined,
+  fundingSource: FundingSource | undefined
+) => {
+  if (!value) return;
+
+  // Cannot determine a funding time without a funding space, as they are space-specific
+  if (!fundingSource) return;
+
+  const sourceTimes = FUNDING_SOURCE_TIMES.find((fst) =>
+    fst.fundingSources.includes(fundingSource)
+  );
+  return sourceTimes.fundingTimes.find((time) =>
+    time.formats.some(
+      (format) => format.toLowerCase() === value.toString().trim().toLowerCase()
+    )
+  )?.value;
+};
