+++ src/entity/decorators/Child/raceValidation.ts
@@ -1,13 +1,8 @@
 import { registerDecorator, ValidationOptions } from 'class-validator';
-import { RACE_FIELDS } from '../../../../client/src/shared/models';
+import { isChildRaceIndicated } from '../../../utils/isChildRaceIndicated';
 import { Child } from '../../Child';
 
 const childRaceNotNullMessage = 'Child race is required';
-
-function isChildRaceIndicated(child: Child) {
-  if (child.raceNotDisclosed) return true;
-  return !RACE_FIELDS.every((field) => !child[field]);
-}
 
 export function ChildRaceIndicated(
   validationOptions?: ValidationOptions
